Moving all ParticleFile tests to new tests/test_particle_file.py file
